move type to top
