+++ metadrive/component/map/waymo_map.py
@@ -4,10 +4,9 @@ from metadrive.component.map.base_map import BaseMap
 from metadrive.component.road_network.edge_road_network import EdgeRoadNetwork
 from metadrive.component.waymo_block.waymo_block import WaymoBlock
 from metadrive.engine.asset_loader import AssetLoader
-from metadrive.scenario.metadrive_type import MetaDriveType
+from metadrive.metadrive_type import MetaDriveType
 from metadrive.utils.waymo_utils.utils import convert_polyline_to_metadrive, read_waymo_data
 from metadrive.utils.waymo_utils.waymo_type import WaymoLaneProperty
-from metadrive.utils.waymo_utils.waymo_type import WaymoRoadLineType, WaymoRoadEdgeType
 
 
 class WaymoMap(BaseMap):
