+++ metadrive/scenario/__init__.py
@@ -1,2 +1,1 @@
-from metadrive.scenario.metadrive_type import MetaDriveType
 from metadrive.scenario.scenario_description import ScenarioDescription
